Make event handler names more specific.
